requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
cmake>=3.28.1
dlib==20.0.0
exceptiongroup==1.3.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
numpy==1.26.4
pillow==10.2.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0

# Compatible con despliegue Render (FastAPI + Dlib + PostgreSQL)